+++ src/edit/holdings/repeatableFields/AdditionalCallNumbersFields.js
@@ -2,19 +2,19 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import CommonAdditionalCallNumbersFields from '../../common/CommonAdditionalCallNumbersFields';
 
-const AdditionalCallNumbersItemLevelFields = (props) => (
+const AdditionalCallNumbersFields = (props) => (
   <CommonAdditionalCallNumbersFields {...props} />
 );
 
-AdditionalCallNumbersItemLevelFields.propTypes = {
+AdditionalCallNumbersFields.propTypes = {
   callNumberTypeOptions: PropTypes.arrayOf(PropTypes.shape({
     label: PropTypes.string,
     value: PropTypes.string,
   })).isRequired,
-  onSwap: PropTypes.func.isRequired,
   canAdd: PropTypes.bool,
   canEdit: PropTypes.bool,
   canDelete: PropTypes.bool,
+  onSwap: PropTypes.func.isRequired,
 };
 
-export default AdditionalCallNumbersItemLevelFields;
+export default AdditionalCallNumbersFields;