+++ CHANGELOG.md
@@ -22,12 +22,13 @@ UIIN-3437.
 * Provide hasCentralTenantPerm arg to useInstancePermissions hook. Fixes UIIN-3444.
 * ECS | No request send when user confirms ownership update of item record. Fixes UIIN-3440.
 * "New request" button is missing from the "Actions" dropdown in the Instance details pane. Fixes UIIN-3445.
-
+* Add additional call numbers to Holdings record. Refs UIIN-3327.
+* Add additional call numbers to Item records. Refs UIIN-3328.
+* 
 ## [13.0.8](https://github.com/folio-org/ui-inventory/tree/v13.0.8) (2025-07-16)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v13.0.7...v13.0.8)
 
 * Always display "Local" in View Source of MARC Holdings records. Refs UIIN-3433.
-* Add additional call numbers to Item records. Refs UIIN-3328.
 
 ## [13.0.7](https://github.com/folio-org/ui-inventory/tree/v13.0.7) (2025-06-30)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v13.0.6...v13.0.7)